+++ tests/testthat/test-pivot-wide.R
@@ -63,6 +63,14 @@ test_that("data frame columns pivot correctly", {
   expect_equal(out$b$y, 7:8)
 })
 
+test_that("works with data.table and empty key_vars", {
+  dt <- data.table::data.table(n = "a", v = 1)
+  expect_equal(
+    pivot_wider(dt, names_from = n, values_from = v),
+    tibble(a = 1)
+  )
+})
+
 
 # column names -------------------------------------------------------------