+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr (development version)
 
+* `pivot_wider()` works with data.table and empty key variables (@mgirlich, #1066).
+
 * `separate_rows()` works for factor columns (@mgirlich, #1058).
 
 # tidyr 1.1.2